load-data: Load international case data from the European CDC
Change the source for international case data from the JHU dataset to
the European CDC dataset. Given the quality issues that we have seen with
the JHU dataset, and the fact that the ECDC data has reasonable country
codings, use their data for all non-US countries.

The tricky bit is that the ECDC data has daily values, but we expect
cumulative values in the database. I added code to generate cumulative
values from the daily values. Additonally, there is some special case
logic to ensure that the US data is removed and that the data for
Britain has the Id mapped from UK->GB, since that is what we expect.