Allow users to specify a custom calibration date
This adds a new advanced option to specify a custom calibration date
that a user can use to tweak simulation start times.

It is an error to attempt to send a calibration date for which we 
don't have data. 

Still need to make sure a sensible error message reaches the user in
this case.

Fixes #55