models: Add a seperate models-staging.yml to allow for separate versioning.
It has been challening to ensure that changes get deployed correctly in
staging because it requires a number of extra steps around tagging,
publishing and updating the `models.yml` file. The majority of this pain
comes from the fact that we were sharing the `models.yml` file between
production and other environments and therefore couldn't relax our
versioning between envrionments.

This change copies the existing `models.yml` file to
`models-staging.yml` and changes all of the model/connector image URLs
to point at the `master` tag rather than a more specific version tag. It
also adds a switch where this file is loaded by the app to load the
correct version based on the environment. The downside to this approach
is that there will be extra effort and attention needed to keep the
other data in these two files in sync when it changes.